Improved blzzd's ntt implementation + more tests and tools.
Changed ntt_blzzd to avoid 64-bit arithmetic. This works for
us because we use a small modulus q.

Added tools to build tables for ntt_blzzd and for computing the
roots psi and phi.

Added tests for the new code. Added a variant of ntt_blzzd with N=1024
for comparison with poly_ntt_harvey and poly_ntt.